html/rendering/non-replaced-elements/tables/table-cellspacing.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Table cellspacing</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/rendering.html#tables-2">
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<table id="table_0" cellspacing="0">
  <thead>
    <tr>
      <th>Month</th>
      <th>Savings</th>
    </tr>
  </thead>
  <tbody>
    <tr>
      <td>January</td>
      <td>$60</td>
    </tr>
    <tr>
      <td>February</td>
      <td>$80</td>
    </tr>
  </tbody>
  <tfoot>
    <tr>
      <td>Sum</td>
      <td>$140</td>
    </tr>
  </tfoot>
</table>

<table id="table_10" cellspacing="10">
  <thead>
    <tr>
      <th>Month</th>
      <th>Savings</th>
    </tr>
  </thead>
  <tbody>
    <tr>
      <td>January</td>
      <td>$60</td>
    </tr>
    <tr>
      <td>February</td>
      <td>$80</td>
    </tr>
  </tbody>
  <tfoot>
    <tr>
      <td>Sum</td>
      <td>$140</td>
    </tr>
  </tfoot>
</table>

<script>

const spacings = ["0", "10"];
for (let spacing of spacings) {
  test(() => {
    let table = document.getElementById(`table_${spacing}`);
    let th_css = window.getComputedStyle(table, null).getPropertyValue("border-spacing");
    assert_equals(th_css, `${spacing}px ${spacing}px`);
  }, `table cellspacing attribute ${spacing}px is correct`)
}

</script>
